### 2025-06-06 #8363 @editorialbot
[PRE REVIEW]: EAVA: An R package for Expert Algorithm Verbal Autopsy (EAVA) cause of death assignment
**Submitting author:** @emilybrownwilson (<a href="http://orcid.org/0000-0001-7487-3371">Emily Wilson</a>)
**Repository:** https://github.com/emilybrownwilson/EAVA
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
