### 2025-06-06 #8363 @samhforbes
Hi @emilybrownwilson thanks for submitting to JOSS. the software will need an open source licence attached to it for us to proceed here.
