### 2025-06-10 818af5f
Add DOIs for openVA and CrossVA; edit spacing of Appendix 2
