+++ paper/paper.bib
@@ -85,7 +85,8 @@
 	year={2024},
 	title={openVA: Automated Method for Verbal Autopsy},
 	journal={https://CRAN.R-project.org/package=openVA},
-	keywords={Package Version 1.1.2}
+	keywords={Package Version 1.1.2},
+	doi={10.32614/CRAN.package.openVA}
 }
 
 @article{transfer,
@@ -183,7 +184,8 @@
 	year={2021},
 	title={CrossVA: Verbal Autopsy Data Transformation for InSilicoVA and InterVA5 Algorithms.},
 	journal={https://CRAN.R-project.org/package=CrossVA},
-	keywords={Package Version 1.0.0}
+	keywords={Package Version 1.0.0},
+	doi={10.32614/cran.package.crossva}
 }
 
 @misc{InterVA,
@@ -205,7 +207,8 @@
 	author={Emily B. Wilson and Henry Kalter and Abhi Datta and Sandipan Pramanik and Robert E. Black},
 	year={2025},
 	title={EAVA: Deterministic Verbal Autopsy Coding with Expert Algorithm Verbal Autopsy},
-	journal={https://CRAN.R-project.org/package=EAVA}
+	journal={https://CRAN.R-project.org/package=EAVA},
+	doi={10.32614/cran.package.eava}
 }
 
 
