### 2025-06-10 6c5a473
Add GPL-2 License
