### 2025-06-10 #8363 @emilybrownwilson
Hi @samhforbes thank you for your review. I've attached the a GPL-2 license.
