### 2025-08-19 #8363 @emilybrownwilson
Hi @samhforbes - I am not sure of the JOSS process after a reviewer declines. Can I offer some names of people familiar with this field of study who may be available? If so, I'm assuming one needs a GitHub account to be a JOSS reviewer? Many thanks.
