### 2025-09-09 #8363 @emilybrownwilson
@samhforbes - what are the next steps for securing a JOSS reviewer for this paper? I can offer a few names if that would help. Please let me know.
