### 2025-11-06 #8363 @goldingn
Hi @jarathomas and @richardli,

Would you both be available to review this R package for analysing verbal autopsies for JOSS?

If you aren't familiar, JOSS reviews are fun and pretty easy and take place interactively on GitHub via checklists and issues. Here's an [example](https://github.com/openjournals/joss-reviews/issues/8445), and here's the [manual](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
