### 2025-11-07 #8363 @jarathomas
Hello,

yeah, I can review.

Regards,
Jason

On Wed, Nov 5, 2025 at 9:15 PM Nick Golding ***@***.***>
wrote:

> *goldingn* left a comment (openjournals/joss-reviews#8363)
> <https://github.com/openjournals/joss-reviews/issues/8363#issuecomment-3494550077>
>
> Hi @jarathomas <https://github.com/jarathomas> and @richardli
> <https://github.com/richardli>,
>
> Would you both be available to review this R package for analysing verbal
> autopsies for JOSS?
>
> If you aren't familiar, JOSS reviews are fun and pretty easy and take
> place interactively on GitHub via checklists and issues. Here's an example
> <https://github.com/openjournals/joss-reviews/issues/8445>, and here's
> the manual
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>
>
