### 2025-12-27 #8363 @derekmeyer37
Hi, Apologies for the late response. I’ll be on vacation for the next few weeks and will be busy with work upon my return. I won’t be able to review this package. Best,DerekOn Dec 19, 2025, at 12:02 AM, Nick Golding ***@***.***> wrote:﻿goldingn left a comment (openjournals/joss-reviews#8363)
Hi @robchallen, @derekmeyer37 would either of you be interested in reviewing this R package submission for JOSS?
