### 2026-01-19 #9778 @robchallen
- Contributions from archived CrossVA not acknowldged & incompatible license used
  - [X] Package methods used in published analyses by current research group
  - [ ] Evidence of adoption by other research groups
  - [ ] Integration into established research workflows or pipelines
  - [X] Demonstrated performance improvements or novel capabilities compared to alternatives
  - [ ] Clear potential for reuse and citation by other researchers in the domain
  - [X] Packaged appropriately according to common community standards
  - [ ] Designed for maintainable extension
  - N.B. single-function packages are not acceptable. This package is 2 functions but one of them is purely data reformatting (and based on previous package).
  - N.B. The EAVA package is an upstream dependency of the openVA package and its functionality is exposed there so there is some suggestion that the methods are more widely adopted but this is not evidenced in this paper.

  - Major functional development present in initial commit
  - No new functionality added after that point just CRAN submission fixes
  - No response to issues raised in Aug 2025
  - No
  - No - Single-author project with no other evidence of community engagement.
  - [X] License: OSI-approved license (required)
  - [ ] Documentation: README, installation instructions,
  - [X] usage examples, API documentation
  - [ ] Quality assurance: Automated tests, continuous integration, and/or documented verification processes
  - [X] Releases: Tagged versions or formal release process
  - [ ] Community pathways: Clear contribution guidelines and support channels
  - Missing critical elements.

  - process not documented but released on CRAN so installation is straightforward.
  - Functions work but no ground truth to verify against. Not able to confirm accuracy of result. Some evidence from Github Issues that functionality is incomplete or defective.

  - No installation instructions provided but package is on CRAN and dependencies automatically resolved.
  - single example provided but expected result not documented.
  - Insufficient documentation of data formats.
  - No. Unaddressed Issues on Github suggesting functionality problems.
  - No.

  - Partially addressed, no clear rationales for build-vs-contribute to openVA which maybe also could have provided a standard data format.
  - Section present but does not explain either rationale behind methods or point to explanation or validation of the algortihm.
  - This is a post-2026 section I think so probably not needed
  - Not present
  - This is a post-2026 section I think so probably not needed
  - Not present
  - This is a post-2026 section I think so probably not needed
