### 2026-04-27 9e26baa
Bug fixes from robust testing (DevPkg-Fun) repo; confirm EAVA package outputs correspond to published results
