+++ R/odk2EAVA.R
@@ -186,7 +186,7 @@ odk2EAVA <- function(odk, id_col) {
   warnZeroMatch <- which(sapply(indexData, length) == 0)
   if (length(warnZeroMatch) > 0) {
     message(paste("Expecting indicator(s) with name(s): ", whoNames[unique(warnZeroMatch)],
-              sep = ""), sep = "\n")
+                  sep = ""), sep = "\n")
     stop("Problem with data: please add above columns to your data frame")
   }
   numNA <- 0
@@ -208,7 +208,7 @@ odk2EAVA <- function(odk, id_col) {
     dups <- lapply(indexData, function(x) length(x) > 1)
     tmpNames <- whoNames[qYesNo]
     message(paste("Duplicate column names containing:", tmpNames[unlist(dups)],
-              sep = " "), sep = "\n")
+                  sep = " "), sep = "\n")
     stop("Problem with data: please remove or rename one of the duplicate columns.")
   }
   iv5Out[, qYesNo] <- as.matrix(odk[, indexData])
@@ -454,6 +454,7 @@ odk2EAVA <- function(odk, id_col) {
   indexData <- which(stri_endswith_fixed(odkNames, whoNames[41]))
   iv5Out[odk[, indexData] > 5, 41] <- "y"
   iv5Out[odk[, indexData] <= 5, 41] <- "n"
+  iv5Out[odk[, indexData] == 98, 41] <- "n"
   iv5Out[odk[, indexData] == 99, 41] <- "n"
   iv5Out[odk[, indexData] == 88, 41] <- "n"
   indexData <- which(stri_endswith_fixed(odkNames, whoNames[43]))
@@ -970,12 +971,13 @@ odk2EAVA <- function(odk, id_col) {
   if (length(indexNA) > 0) {
     warning("NA's included in output", call. = FALSE)
     message(paste("odk2openVA produced NA's in the following columns",
-              " (this may cause errors with openVA)", sep = ""),
-        sep = "\n")
+                  " (this may cause errors with openVA)", sep = ""),
+            sep = "\n")
     message(paste(iv5Names[indexNA], " Probably associated with WHO column containing: ",
-              whoNames[indexNA], sep = ""), sep = "\n")
+                  whoNames[indexNA], sep = ""), sep = "\n")
   }
-  indexID <- which(stri_endswith_fixed(odkNames, id_col))
+  # indexID <- which(stri_endswith_fixed(odkNames, id_col))
+  indexID <- which(stri_endswith_fixed(odkNames, tolower(id_col)))
   # indexID <- odkNames[4]
   if (length(indexID)) {
     iv5Out <- cbind(as.character(odk[, indexID]), iv5Out)
@@ -1013,11 +1015,11 @@ odk2EAVA <- function(odk, id_col) {
 
   odkNames <- tolower(names(odk))
   whoNames_add <- c("Id10183","Id10167","Id10173","Id10161","Id10250","Id10120","Id10182","Id10148","Id10234",
-                    "Id10126","Id10446","Id10151","Id10154")
+                    "Id10126","Id10446","Id10151","Id10154","Id10184_b")
   whoNames_add <- tolower(whoNames_add)
   eavaNames <- c("i183b","fb_day0","i173b","i167c","i161b","i250b","i120c","i182d","i148d","i234c",
[…]
