+++ R/codEAVA.R
@@ -19,14 +19,15 @@
 codEAVA <- function(df,age_group){
   if(age_group=="neonate"){
     data <- df
-    data$Stillbirth <- ifelse(data$i104o %in% c("n",".") & data$i109o %in% c("n",".") & data$i110o %in% c("n","."),1,0)
+    # data$Stillbirth <- ifelse(data$i104o %in% c("n",".") & data$i109o %in% c("n",".") & data$i110o %in% c("n","."),1,0)
+    data$Stillbirth <- ifelse(data$i104o %in% c("n") & data$i109o %in% c("n") & data$i110o %in% c("n"),1,0)
     data <- data[data$Stillbirth!=1,]
 
     data <- subset(data, age <28)
     dim(data)
 
     ## |++| neonatal tetanus1 |++|
-    data$nnt1 <- ifelse((data$age > 2 & data$age < 27) & data$i219o == "y" & ((data$i271o == "y" & data$i273o == "y") | (data$i106a == "y" & data$i107o == "y")),1,2)
+    data$nnt1 <- ifelse((data$age > 2 & data$age < 27) & data$i219o == "y" & ((data$i271o == "y" & data$i273o == "y") | (data$i106a == "n" & data$i107o == "y")),1,2)
     table(data$nnt1, exclude = NULL)
 
     ## | + congenital malformation2 + |
@@ -58,7 +59,7 @@ codEAVA <- function(df,age_group){
     table(data$meningitis451_nonnt1, exclude = NULL)
 
     # * |++| diarrhea8 |++| (for NNs: acute diarrhea with >4 stools on worst day);
-    data$diarrhea8 <- ifelse(data$i181o == "y" & data$i183b=="y", 1,2)
+    data$diarrhea8 <- ifelse(data$i181o == "y" & data$i183b == "y", 1,2)
     table(data$diarrhea8, exclude = NULL)
 
     # * |++| pneumonia157 |++|;
@@ -70,7 +71,7 @@ codEAVA <- function(df,age_group){
 
     data$pneumo157signs <- data$pneumo157sign2 + data$pneumo157sign3 + data$pneumo157sign5
 
-    data$pneumonia157 <- ifelse((data$i166o == "y" & data$i167c == "y") | data$i159o %in% "y" & data$i161b == "y" & data$pneumo157signs>1, 1, 2)
+    data$pneumonia157 <- ifelse((data$i166o == "y" & data$i167c == "y" | data$i159o %in% "y" & data$i161b == "y") & data$pneumo157signs>1, 1, 2)
     table(data$pneumonia157, exclude = NULL)
 
     # * |++| sepsisfvr |++|;
@@ -96,6 +97,10 @@ codEAVA <- function(df,age_group){
     # * |++| sepsisfvr2_2 without neonatal tetanus |++|;
     data$sepsisfvr2_2_nonnt1 <- ifelse(data$sepsisfvr2_2 == 1 & data$nnt1 == 2, 1, 2)
     table(data$sepsisfvr2_2_nonnt1, exclude = NULL)
+
+    # * |++| possible diarrhea |++|;
+    data$possiblediar8_8 <- ifelse(data$i181o %in% "y" & data$sepsisfvr2_2 %in% 1 & data$diarrhea8 %in% 2, 1,2)
+    table(data$possiblediar8_8, exclude = NULL)
 
     # * |++| possible pneumonia |++|;
     data$possiblepneumonia9 <- ifelse(data$i159o == "y" & data$sepsisfvr2_2 == 1 & data$pneumonia157 == 2, 1, 2)
@@ -146,7 +151,7 @@ codEAVA <- function(df,age_group){
     table(data$malnutrition1, exclude = NULL)
 
[…]
