### 2026-04-27 9ca9d51
License changed from GPL-2 to GPL-3; improve documenation of CrossVA in odk2EAVA.R file; employ NEWS.md file for tracking updates
