+++ cran-comments.md
@@ -1,13 +1,15 @@
+## Resubmission 20260427
+This resubmission is an update to version 1.0.0 to 1.1.0 reflecting improvements from JOSS paper peer review
+* Changes detailed in NEWS.md file
+
 ## Resubmission
 This is the third submission. In this version I have:
-
 * Referenced an archived R package, which odk2EAVA() builds upon, in description field of DESCRIPTION file
 * Removed quotes around Expert Algorithm Verbal Autopsy for consistency
 * Added small executable examples in man/...Rd files to illustrate function use and enable testing
 * Changed cat() to message() in R/odk2EAVA.R
 
 ## Changes in the previous resubmission
-
 * Removed Author field that differed from Authors@R 
 * Added imports to NAMESPACE
 * Defined globalVariable age in zzz.R
