+++ LICENSE.md
@@ -1,248 +1,124 @@
-GNU General Public License version 2
+GNU General Public License version 3
 ====================================
-https://opensource.org/license/gpl-2-0
-Version 2.0, 10 June 2025
+
+Version 3.0, 5 September 2007
 
 Preamble
 ========
 
-The licenses for most software are designed to take away your freedom to share and change it. 
-By contrast, the GNU General Public License is intended to guarantee your freedom to share and 
-change free software–to make sure the software is free for all its users. This General Public 
-License applies to most of the Free Software Foundation’s software and to any other program 
-whose authors commit to using it. (Some other Free Software Foundation software is covered by 
-the GNU Library General Public License instead.) You can apply it to your programs, too.
-
-When we speak of free software, we are referring to freedom, not price. Our General Public Licenses 
-are designed to make sure that you have the freedom to distribute copies of free software (and 
-charge for this service if you wish), that you receive source code or can get it if you want it, 
-that you can change the software or use pieces of it in new free programs; and that you know you 
-can do these things.
-
-To protect your rights, we need to make restrictions that forbid anyone to deny you these rights 
-or to ask you to surrender the rights. These restrictions translate to certain responsibilities 
-for you if you distribute copies of the software, or if you modify it.
-
-For example, if you distribute copies of such a program, whether gratis or for a fee, you must 
-give the recipients all the rights that you have. You must make sure that they, too, receive or 
-can get the source code. And you must show them these terms so they know their rights.
-
-We protect your rights with two steps: (1) copyright the software, and (2) offer you this license 
-which gives you legal permission to copy, distribute and/or modify the software.
-
-Also, for each author’s protection and ours, we want to make certain that everyone understands that 
-there is no warranty for this free software. If the software is modified by someone else and passed 
-on, we want its recipients to know that what they have is not the original, so that any problems 
-introduced by others will not reflect on the original authors’ reputations.
-
-Finally, any free program is threatened constantly by software patents. We wish to avoid the danger 
-that redistributors of a free program will individually obtain patent licenses, in effect making the 
-program proprietary. To prevent this, we have made it clear that any patent must be licensed for 
-everyone’s free use or not licensed at all.
-
+The GNU General Public License is a free, copyleft license for software and other kinds of works.
+The licenses for most software and other practical works are designed to take away your freedom to share and change the works. By contrast, the GNU General Public License is intended to guarantee your freedom to share and change all versions of a program–to make sure it remains free software for all its users. We, the Free Software Foundation, use the GNU General Public License for most of our software; it applies also to any other work released this way by its authors. You can apply it to your programs, too.
+When we speak of free software, we are referring to freedom, not price. Our General Public Licenses are designed to make sure that you have the freedom to distribute copies of free software (and charge for them if you wish), that you receive source code or can get it if you want it, that you can change the software or use pieces of it in new free programs, and that you know you can do these things.
+To protect your rights, we need to prevent others from denying you these rights or asking you to surrender the rights. Therefore, you have certain responsibilities if you distribute copies of the software, or if you modify it: responsibilities to respect the freedom of others.
+For example, if you distribute copies of such a program, whether gratis or for a fee, you must pass on to the recipients the same freedoms that you received. You must make sure that they, too, receive or can get the source code. And you must show them these terms so they know their rights.
+Developers that use the GNU GPL protect your rights with two steps: (1) assert copyright on the software, and (2) offer you this License giving you legal permission to copy, distribute and/or modify it.
[…]
