+++ NEWS.md
@@ -1,9 +1,7 @@
+# EAVA 1.1.0
+* Bug fixes to odk2EAVA and codEAVA functions - detailed in DevPkg-Fun repository
+* License changed from GPL-2 to GPL-3 to be compatible code bundled from CrossVA
+* Improve documentation of CrossVA in odk2EAVA.R file
+
 # EAVA 1.0.0
-
-# EAVA 2.0.0
-
-# EAVA 1.0.0
-
-# EAVA 0.1.0
-
 * Initial CRAN submission.
