+++ DESCRIPTION
@@ -35,13 +35,13 @@ Authors@R: c(
   )
 Maintainer: Emily Wilson <wilsonem@gmail.com>
 Description: Expert Algorithm Verbal Autopsy assigns causes of death to 2016 WHO Verbal Autopsy Questionnaire data. odk2EAVA() converts data to a standard input format for cause of death determination building on the work of Thomas (2021) <https://cran.r-project.org/src/contrib/Archive/CrossVA/>. codEAVA() uses the presence and absence of signs and symptoms reported in the Verbal Autopsy interview to diagnose common causes of death. A deterministic algorithm assigns a single cause of death to each Verbal Autopsy interview record using a hierarchy of all common causes for neonates or children 1 to 59 months of age.
-License: GPL-2
+License: GPL-3
 Encoding: UTF-8
 LazyData: true
 Imports: 
     stringi,
     stringr
-RoxygenNote: 7.3.2
+RoxygenNote: 7.3.3
 BugReports: https://github.com/emilybrownwilson/EAVA/issues
 Depends: 
     R (>= 2.10)
