### 2026-04-27 90d598f
Add column descriptions and source references for example data and the 2016 WHO verbal autopsy instrument in data.R
