+++ NEWS.md
@@ -2,6 +2,7 @@
 * Bug fixes to odk2EAVA and codEAVA functions - detailed in DevPkg-Fun repository
 * License changed from GPL-2 to GPL-3 to be compatible code bundled from CrossVA
 * Improve documentation of CrossVA in odk2EAVA.R file
+* Add column explanation and sources in data.R
 
 # EAVA 1.0.0
 * Initial CRAN submission.
