### 2026-04-27 874e41f
Add README.Rmd, CONTRIBUTING.md
