+++ NEWS.md
@@ -3,6 +3,9 @@
 * License changed from GPL-2 to GPL-3 to be compatible code bundled from CrossVA
 * Improve documentation of CrossVA in odk2EAVA.R file
 * Add column explanation and sources in data.R
+* Add CONTRIBUTING.md
+* Add README.md
+* Harmonize version changes in NEWS.md
 
 # EAVA 1.0.0
 * Initial CRAN submission.
