### 2026-04-28 b9bf7a9
Add more input file information for odk2EAVA and create data/ccva_codebook.rda
