### 2026-04-28 e18d847
Update NEWS.md
