+++ man/data_public.Rd
@@ -3,9 +3,9 @@
 \docType{data}
 \name{data_public}
 \alias{data_public}
-\title{Example WHO 2016 Verbal Autopsy Dataset from the Countrywide Mortality Surveillance for Action (COMSA) project in Mozambique}
+\title{Example 2016 WHO Verbal Autopsy Dataset from the Countrywide Mortality Surveillance for Action (COMSA) project in Mozambique}
 \format{
-`data_public` is example data frame with 10 rows and 511 columns comprising verbal autopsy questionnaire modules
+`data_public` is an example data frame with 10 rows and 511 columns comprising verbal autopsy questionnaire modules
 \describe{
   \item{IDs, geolocation, weighting}{comsa_id-adj_wt2007}
   \item{indetifiers of respondent and interviewer}{Id10007-Id10011}
