### 2026-05-04 7288f4d
Finalize data/ccva_codebook.rda
