+++ NEWS.md
@@ -6,7 +6,8 @@
 * Add CONTRIBUTING.md
 * Add README.md
 * Harmonize version changes in NEWS.md
-* Add more input file information for odk2EAVA and create data/ccva_codebook.rda 
+* Add more input file information for odk2EAVA 
+* Create data/ccva_codebook.rda with R/build_ccva_codebook.R
 
 # EAVA 1.0.0
 * Initial CRAN submission.
