### 2026-08-28 #9778 @jarathomas
Hi @emilybrownwilson, I created a few issues on the EAVA repo.  Once those are resolved, I can check of the rest of the items.
